Updated management service name & introduced the dashboard endpoint
