+++ services/admin/public/login.html
@@ -28,7 +28,7 @@
                     id="loginForm"
                     class="form"
                     method="POST"
-                    action="/login/auth"
+                    action="/admin/login/auth"
                     onsubmit="login(event)"
                 >
                     <span class="form__title">Admin Page</span>
