+++ services/admin/routes/admin/dashboard/index.js
@@ -1,0 +1,9 @@
+const express = require('express');
+
+const router = express.Router();
+
+const dashboard = require('./dashboard');
+
+router.use('/', dashboard);
+
+module.exports = router;
