+++ services/admin/routes/admin/index.js
@@ -3,7 +3,9 @@ const express = require('express');
 const router = express.Router();
 
 const loginEndpoint = require('./login');
+const dashboardEndpoint = require('./dashboard');
 
 router.use('/login', loginEndpoint);
+router.use('/dashboard', dashboardEndpoint);
 
 module.exports = router;
