+++ services/admin/routes/index.js
@@ -1,0 +1,9 @@
+const express = require('express');
+
+const router = express.Router();
+
+const adminEndpoint = require('./admin');
+
+router.use('/admin', adminEndpoint);
+
+module.exports = router;
